refactor NFS views with Mixin and simplify. #876
